@@ -8,11 +8,11 @@ interface SerializerInterface
      * @param mixed $data
      * @return string
      */
-    public function serialize($data): string;
+    public function serialize(mixed $data): string;
 
     /**
      * @param string $serialized
      * @return mixed
      */
-    public function unserialize(string $serialized);
+    public function unserialize(string $serialized): mixed;
 }
